Reorganize for typescript
